+++ src/RcppExports.cpp
@@ -27,16 +27,18 @@ BEGIN_RCPP
 END_RCPP
 }
 
-RcppExport SEXP read_tags_c(SEXP, SEXP);
-RcppExport SEXP read_tif_c(SEXP);
-RcppExport SEXP write_tif_c(SEXP, SEXP, SEXP, SEXP, SEXP);
+RcppExport SEXP count_directories_C(SEXP);
+RcppExport SEXP read_tags_C(SEXP, SEXP);
+RcppExport SEXP read_tif_C(SEXP);
+RcppExport SEXP write_tif_C(SEXP, SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
     {"_ijtiff_dims_cpp", (DL_FUNC) &_ijtiff_dims_cpp, 1},
     {"_ijtiff_float_max", (DL_FUNC) &_ijtiff_float_max, 0},
-    {"read_tags_c",       (DL_FUNC) &read_tags_c,       2},
-    {"read_tif_c",        (DL_FUNC) &read_tif_c,        1},
-    {"write_tif_c",       (DL_FUNC) &write_tif_c,       5},
+    {"count_directories_C", (DL_FUNC) &count_directories_C, 1},
+    {"read_tags_C",         (DL_FUNC) &read_tags_C,         2},
+    {"read_tif_C",          (DL_FUNC) &read_tif_C,          1},
+    {"write_tif_C",         (DL_FUNC) &write_tif_C,         5},
     {NULL, NULL, 0}
 };